Fix syntax warning by checking for equality rather than identity on the integer zero.
This works both ways because, specifically in Python, the integers from 0 to 255are stored once and passed by reference rather than by value, but the value of having the englishy 'is not' in here is outweighed by the value of a maintainer not needing to know that. A maintainer is more likely to recognize the '!=' syntax than to know this weird thing about the Python compiler.