Miscellaneous fixes for the firefox-ci export job (#230)
* fix(fxci): make bigquery table names configurable

The tables are defined and versioned in bigquery-etl. Make the names
configurable so we don't need to rebuild the image when they change.

* fix(fxci): don't assume the last run is the one that generated the pulse event

While testing, I discovered that it's possible for a pulse event to
contain runs later than the one that generated the event. So it's not
safe to simply grab the last run in the list.

Instead we should use the run from the index in the top-level `runId` field.

* fix(fxci): skip tasks that hit 'deadline-exceeded'