feat(fxci): add logic to automatically create tables with correct schemas if they don't exist
This ended up being a bit more complicated than I wanted, but since I
had all the types of the records defined in the dataclass, I didn't want
to hard code them all just for a BigQuery schema. So instead this uses
the `generate_schema` function to automatically derive the table schema
from the `Record` classes.

Was it worth doing this way? Probably not. But it works well, so I think
it's worth keeping. This code can be re-used for future tables as well.