fix(fxci): change metric export interval calculation to do one day at a time
The worker uptime data and worker cost data need to be as synced as
possible. For example, if the uptime data lags half a day behind the
cost data, that will throw off the cost calculation for tasks.

The worker cost export runs in the `bigquery-etl` repo on a daily
schedule. It only processes data for the last complete day we have
(e.g yesterday).

Therefore, this patch makes the worker uptime export do the same thing.
In this way, we should be able to guarantee that uptime and cost data
are about as in sync as possible.