refactor(fxci): tie bq loader instances to specific tables
This removes some of the smarts in the BigQueryLoader class. Rather than
accepting a mish mash of records that can be from any table, we now tie
loader instances to specific tables.

This makes a tiny bit more work for callers of this class, but makes the
overall architecture of the ETL much simpler.