Support a --date parameter in the fxci metric export (#274)
* refactor(fxci): tie bq loader instances to specific tables

This removes some of the smarts in the BigQueryLoader class. Rather than
accepting a mish mash of records that can be from any table, we now tie
loader instances to specific tables.

This makes a tiny bit more work for callers of this class, but makes the
overall architecture of the ETL much simpler.

* feat(fxci): use a --date flag to derive metric export time interval

This improves the metric export such that it receives a date as an input
parameter rather than tracking start/end times explicitly. The dates are
idempotent, meaning if you re-run a date, all existing records will
first be deleted before inserting the new ones.

This allows us to backfill data from past dates, or re-generate
corrupted data. It also makes the logic for the metric export much
simpler.

Bug: 1912067

* chore: remove dependency on pytz

This can all be accomplished with the stdlib as of Python 3.9