Detect and correct wraparound for negative PPA conversion counts
Differential privacy in [Prio][1] can result in negative conversion
counts, which then wrap around to large numbers as all operations are
performed modulo a large prime number.

This change hard-codes the prime number used for the Prio3SumVec data
type (currently the only data type we use for DAP reports) to detect and
correct this behavior, such that we can generate human-friendly reports
for advertisers.

Namely, given the modulus P is a 128-bit prime, any value N in the
histogram that is greater than 2^127 is replaced with (N-P).

JIRA: [AE-553][2]

[1]: https://github.com/divviup/libprio-rs/
[2]: https://mozilla-hub.atlassian.net/browse/AE-553