Support importing CrUX data into webcompat kb (#298)
* Move bugzilla code to a submodule

* Refactor argument handling

Use `--bq-project` and `--bq-kb-dataset` as aliasses for
`--bq_project_id` and `--bq_dataset_id`, but make the old values still
work to avoid needing to update prod yet.

Default to a test dataset for the kb datset.

Move bugzilla-specific arguments to the bugzilla module.

* Refactor to support running multiple ETL jobs

Modules are just explicitly defined in a dictionary in main.

Move creating the client to the main function, and pass it in
to the main() function for each module.

* Disallow untyped defs in mypy configuration

* Add webcompat-kb job for updating CrUX data

This updates our CrUX ranking data from the upstream tables. The
upstream data is updated once a month with the data for the previous
month, so we avoid running any updates unless we are missing data
corresponding to the previous month.