Import CrUX data into hosts_min_ranks table
This is a direct view of the specific data we want: global minimum rank,
local minimum rank, and sightline minimum rank. The requirement for a
subquery to select the latest CrUX import makes it hard to write as
a materialized view, but since we only need to update it when the
CrUX data itself updates, adding it to the ETL makes sense.