Store task tags in a separate table (#329)
* fxci: add config.dev.toml to .gitignore

I've been using this for local testing, want to make sure I don't
accidentally commit secrets.

* fxci: fix bug in generate_schema

The logic had `Optional` and `list` in the same if/elif/else block. This
meant that, e.g `Optional[list[str]]` wasn't getting handled properly.
Of course that is a valid type!

Do `list` handling after `Optional` handling to fix it.

* fxci: refactor task tags so each tag is a separate column

I made a mistake with the data model. Initially I thought it would be
really nice if we could just add an arbitrary tag to a task and then
immediately query it in the tables / dashboards. So I decided to make
tags a nested column in the tasks table, containing "key" and "value"
columns.

It turns out while this is nice if you just want to query a single tag,
it adds a ton of complexity if you want to be able to filter by multiple
tags at once. This seems obvious in hindsight, but I naively thought
there would be some good way of handling this.

Instead I'd like to redo the data model so that each unique tag is
stored in in a distinct column. This will make it way easier to query
multiple tags, with the downside being that we'll need to explicitly
update the table schema anytime we want to start tracking a new tag. I
don't expect we'll need to do this too often, and when we do, it should
be fairly easy. I'll make sure to add documention for this as well.