fix(fxci): bump version in tasks table (#330)
I wrongfully thought that the schema changes to the tasks table that I
made were backwards compatible, but I missed that I was changing it from
a REPEATED STRUCT column to just a STRUCT.

So let's create a new v2 table. I'll try and write a backfill task /
script to repopulate the old data from the v1 table later.