fix(fxci): ensure bigquery tables before consuming pulse events
By instantiating the BigQueryLoader in the on_processing_complete
callback, we don't validate the tables until *after* we drain the pulse
queue. This is bad because if there's an error (like a schema mismatch),
we lose all those events forever.

This way the pulse messages don't get consumed, so the data will still
be there for consumption once the issue is fixed.