Convert bug history import to use search API
This allows us to get the history for many bugs in a single query,
rather than needing one query per bug, for a substantial perf win. The
downside is we always fetch the full history, so for incremental
updates we filter out the older entries.

This also changes a lot of the history code to use datclasses or other
type-friedly APIs, rather than passing around dicts.