Add a job that stores changes to the metric value.
This works by reconstructing historic states of each bug, computing
the score for each state, and then storing all the deltas along with
the kind of bug change that caused the score change.

To compute all the scores using the same logic as the backend, we
create a table with all the historic states, use that to perform the
score calculations, and then remove it. We could perhaps do that
better (or just keep the table permanently, if the data is useful).