Record per-bug changes to the webcompat metric (#334)
* Fix test.sh invocation of mypy

* Put command line argument in correct group

* Add --pdb flag

This is for local development and drops into a debugger if there's an
unhandled exception.

* Add a development command to reload test data

Add a webcompat-create-test-dataset command that clones bugzilla data
from the main dataset to a {prod name}_test dataset which can be used
for testing local changes. This is significnatly faster than
re-running the main import.

* Improve logging message

* Add a flag to bugzilla job to recreate history from scratch

* Don't store history fetch completed on the class instance

This makes it easier to reason locally about the places where the
variable can change.

* Store creator and user_story_raw on bugs

These are needed for computing the score history. In particular the unparsed user story
is required to apply diffs from history.

* Store change history for url and user_story fields

These are part of the score computation, so we need to store changes in order to accurately compute
historic scores

* Log warning if bug history isn't going to be written

* Convert bug history import to use search API

This allows us to get the history for many bugs in a single query,
rather than needing one query per bug, for a substantial perf win. The
downside is we always fetch the full history, so for incremental
updates we filter out the older entries.

This also changes a lot of the history code to use datclasses or other
type-friedly APIs, rather than passing around dicts.

* Fix bugzilla tests for API changes

* Add a job that stores changes to the metric value.

This works by reconstructing historic states of each bug, computing
the score for each state, and then storing all the deltas along with
the kind of bug change that caused the score change.

To compute all the scores using the same logic as the backend, we
create a table with all the historic states, use that to perform the
score calculations, and then remove it. We could perhaps do that
better (or just keep the table permanently, if the data is useful).