Create a bigquery wrapper and use it everywhere
This centalises a lot of our common patterns in one place.

Each job gets a wrapper with the dataset approproately configured for
the job. The wrapper also knows about the write flag, which it can use
to avoid writes when creating tables or loading data, which is the
common case, but it isn't able to prevent general queries from
performing inserts.