Refactor bugzilla import (#352)
The previous import was getting rather hard to understand, because it
was mixing IO with bug grouping logic in a way that made it difficult
to tell which bugs ended up in which groups.

In the new version, we simply load all the bugs that we are interested
in into a single large dict at the start. We then do all the grouping
and filtering we need in Python code, before finally writing all the
data to bigquery.

Unlike bugs themselves, bug history is a bit more complicated because we allow
incremental updates. This is currently grouped into a single class
that only covers history updating, but some additional work would be
required to make it match the pure input/processing/output flow that
we have for bugs themselves.

In addition this change tries to used defined types wherever possible,
and avoids using untyped dictionaries except at the edges. This makes
it much easier to catch a class of errors using mypy.