Create a bigquery wrapper and use it everywhere (#349)
* Create a bigquery wrapper and use it everywhere

This centalises a lot of our common patterns in one place.

Each job gets a wrapper with the dataset approproately configured for
the job. The wrapper also knows about the write flag, which it can use
to avoid writes when creating tables or loading data, which is the
common case, but it isn't able to prevent general queries from
performing inserts.

* Fix bug leading to duplication of metric change entries (#351)

* Fix bug leading to duplication of metric change entries

The underlying bug was a broken check for the bug being new, which
meant we considered old bugs with changes to always be new.

In addition add a verification step that will reject any changes where
we already have an entry for the same bug with the same timestamp.

* Update jobs/webcompat-kb/webcompat_kb/metric_changes.py

Co-authored-by: Dennis Schubert <mail@dennis-schubert.de>

---------

Co-authored-by: Dennis Schubert <mail@dennis-schubert.de>

---------

Co-authored-by: Dennis Schubert <mail@dennis-schubert.de>