Handle missing dependencies (#353)
Sometimes we can't access dependent bugs because they're sec bugs or
something else we don't have access to. In this case we need to:

* Not keep trying to fetch the bug forever, even though it looks like
a missing bug.
* Ensure that we don't fail later when the bug isn't present

The other approach here would be to use a placeholder entry for the
missing bug. That would have the advantage that it would maintain the
same invariants we previously had, and reduce data loss, but at the
cost of having to create useful fake data for the missing bugs.