Switch to insert_rows from insert_rows_json
That does the type conversions we want.

For write_table we are still using `load_table_from_json` so add a
read-only Json type, and specify that as the type for rows. This means
that we get a type error if we make a mistake like trying to pass in a
datetime object rather than a string-formatted datetime.