Reduce the number of separate bugzilla queries
This avoids fetching the same bug multiple times in some cases.

We use one query for all bugs in the Web Compatibility component and
one query for all bugs outside Web Compatibility. This also unifomly
enforces the limit that we only import bugs that are open, or which
were resolved after Jan 1st 2020. As a result we stop importing some
older Interventions bugs, but we weren't using those for anything
anyway. We also stop importing parity- bugs in the Web Compatibility
product, but in general we don't expect such bugs to be important.