Require getting a lock before running the ETL
This is a bad idea

Sometimes we want to run specific ETL jobs locally e.g. when updating
the scoring logic. When doing this we don't want the ETL itself
updating data, as a lot of the logic assumes that we don't have
concurrent writes.

This approach uses a very simple implementation of a lock table:
essentially a table that has zero or one rows at a given time. If
there's an existing row the lock is taken and the process either exits
or polls the lock until it can be acquired.