Import web features data into bigquery (#350)
This uses the web features releases from
https://github.com/web-platform-dx/web-features/ and reads them into
the following tables:

releases: a list of all released versions of the web features dataset
browsers: a list of all browser names and versions that appear
anywhere in any web features release
features: a table listing each feature in each release of the
data. Browser support is encoded as a repeated record in the 'support' column.