Metric rescore (#359)
* Allow the concept of non-default jobs

This is to support running some jobs locally rather than in the deployed ETL.

* Make inserting rows to the import_runs table less likely to fail

In particular create the table if it doesn't exist, and depend on
the BQ data format conversion rather than doing it by hand.

* Add methods to get tables and routines to BigQuery wrapper.

* Store old and new scores in metric_changes table

* Move to using Decimal everywhere to represent scores

This maps to the NUMERIC type in BigQuery, and avoids precision loss
when converting to and fron floats.

* Change the way we recreate the metric changes table

Instead of using the tactic of deleting the table and later writing a
new one, write with the write_mode of WRITE_TRUNCATE.

* Remove recreate argument from ensure_table

* Update README to better describe local development

* Add a non-default job for rescoring the metric

This is intended to me run manually, along with the bugzilla job, the
metric job, and the metric_changes job.

It depends on having a view with the old scoring logic alongside the
existing scored_site_reports view. The name of this view is passed in
using --metric-rescore-new-scored-site-reports

In addition if any of the score computation routines are changed these
must be passed in on the command line using --metric-rescore-update-routine

It does three things:
* Computes: the total metric scores with both the old and new views,
and stores them in a webcompat_topline_metric_rescores table.
* Computes: the score change for all open bugs that have a different
score before and after the rescore, and adds these to the
webcompat_topline_metric_changes table with the reason set to
"rescore". In addition entries with zero score change are added for
any bugs that moved into or out of the sightline or global_1000 states
due to the rescoring with reason set to (one or more of)
"sightline-removed", "sightline-added", "global-1000-removed",
"global-1000-added".
* Updates routines and view definitions so that the existing
scored_site_reports view is saved as
scored_site_reports_before_[datetime] and the scored_site_reports is
updated with the query from the new scored site reports view. If any
routines are changed they are archived and updated in a similar way,
and the SQL is also updated so the archived view points to the
archived routine, and the live view points to the canonical name of
the routine (this is done using string subsitution so is not super robust).

* Update web_features job for removed recreate argument