Only re-import updated bugs (#362)
This modifies the bugzilla queries, adding a condition to only return
bugs that have changed since the last successful update. As a result
instead of reading several thousand bugs and taking 10+ minutes to
run, this should typically import fewer than 10 new bugs each run, and
cuts the runtime to < 1 minute.

The logic still assumes that we have all bugs available locally, and
that we can overwrite the entire bugs table each time, so we read all
the existing bug data from BigQuery (which is much faster) and only
overwrite the updated bugs.

To make this easier we also store a couple of new fields in the bugs
table, so that we have see_also and last_change_time data available
when reading back from BigQuery.