Add a site-ranks-update manual job
This is designed to make it easier to automatically update to a new
version of site rank data, on the basis that all data is represented
as a yyyymm integer.

The job automatcially creates a new version of the routine containing
the current yyyymm value that we're using; updates a copy of
`scored_site_reports` to use the new version, and then runs the
rescoring logic to record the metric changes resulting from the change
in the site rank data.