Import tranco data into webcompat dataset
This is imported into a new tranco_imported dataset by default. We
currently import the 1M top subdomains list.

By default a tranco import is performed whenever there is a new
release of CrUX data. The destination table therefore has a similar
yyyymm key and similar partitioning.

This PR doesn't yet change the way that we compute host_min_ranks to
include the tranco data, so it doesn't yet affect the scoring in any way.