Slightly refactor metrics code
Define all the metrics in one place in the code.

The Metric class defines the kinds of metrics, and the field on
`scored_site_reports` that is true for rows in that metric.

The MetricType class defines all the variants of each metric (bug
count, needs_diagnosis, not_supported, total), the conditional applied
to `scored_site_reports` to determine if a row is part of that
variant, and the aggregation function used to create the final value
of the metric variant.

When defining queries operating over all metrics and variants write
them as loops over lists of all Metric and MetricType instances.