Distinguish between metric types we record in different contexts
For some reason that's not entirely clear we record the platform_score
metric type only in the webcompat_topline_metric_*_history tables.

In lieu of fixing the underlying problem, give each metric type a
`context` field which is a set of strings refering to the different
table types in which we can store the metric. Things with the
`history` context type go in the _history tables, things with the
_daily context type go in the _daily tables (and also in the _rescores table).