Add a command for backfilling the metric scores
When we add a new metric type we generally want to backfill the
results we would have got at earlier dates based on the current
data in bugzilla. This won't always match the actual results on those
dates because we're looking at current bug information rather than the
bug as it would have appeared on that date, but it's generally good
enough to provide some historical context.

For this reason add a metric-backfill command that is intended to be
used interactively with a `--metric-backfill-metric` argument to
specify the metric that should be backfilled.