Japan 1000 mobile (#393)
* Update list of webcompat views+tables to sync into _test dataset

* Slightly refactor metrics code

Define all the metrics in one place in the code.

The Metric class defines the kinds of metrics, and the field on
`scored_site_reports` that is true for rows in that metric.

The MetricType class defines all the variants of each metric (bug
count, needs_diagnosis, not_supported, total), the conditional applied
to `scored_site_reports` to determine if a row is part of that
variant, and the aggregation function used to create the final value
of the metric variant.

When defining queries operating over all metrics and variants write
them as loops over lists of all Metric and MetricType instances.

* Add japan_1000_mobile metric

This will ensure the metric value is recorded each day and created in
the test dataset.

* Fix pyproject.toml fields

* Distinguish between metric types we record in different contexts

For some reason that's not entirely clear we record the platform_score
metric type only in the webcompat_topline_metric_*_history tables.

In lieu of fixing the underlying problem, give each metric type a
`context` field which is a set of strings refering to the different
table types in which we can store the metric. Things with the
`history` context type go in the _history tables, things with the
_daily context type go in the _daily tables (and also in the _rescores table).

* Add a command for backfilling the metric scores

When we add a new metric type we generally want to backfill the
results we would have got at earlier dates based on the current
data in bugzilla. This won't always match the actual results on those
dates because we're looking at current bug information rather than the
bug as it would have appeared on that date, but it's generally good
enough to provide some historical context.

For this reason add a metric-backfill command that is intended to be
used interactively with a `--metric-backfill-metric` argument to
specify the metric that should be backfilled.