Add import of Interop issues into webcompat_kb
Currently this is just a simple import that looks for issues labelled
with either `focus-area-proposal` or `investigation-proposal` and
stores a row with the issue number and name of the proposal, along
with any web-features or bugzilla issues linked in either the original
proposal or the web-features bot comment.

Later we will probably want some way to handle proposals being closed
/ accepted / etc. but that can be a followup.

Due to GitHub rate limits, this requires a GitHub token to run correctly.