Update web_features importer for version 3.x (#400)
This adds the possibility of features being moved to a different name,
or split into multiple different features.

We simply write these to specific tables for each kind of change. In
practice moves we can probably handle with a view that automatically
selects the new data for the old name, whereas splits require human
intervention to find the correct new feature to use.