fxci-etl: add and populate a task_definitions table (bug 1989550)
Process task-defined pulse messages to keep a set of new task ids, whose
definitions we then look up in taskcluster.

Add a new object to the backup storage with a list of task ids we have
seen but whose definitions we haven't been able to add.

These new records are added to bigquery in batches of 1000 instead of
5000 to hopefully avoid running into request size limits, since they can
get quite big.