Create a job to deploy views and routines
Each view or routine is defined by a meta.toml file containing the
name of the view or routine and other metadata, and a view.sql or
routine.sql file defining the actual SQL.

The SQL definition is a Jinja template, which has several useful
values and functions:

 * `ref()` for references to other data sets. This allows tracking of
 dependencies.

 * metrics, a dictionary of {metric_name: Metric}

 * metric_types, a list of MetricType objects

The routines are sorted topologically for deployment (so that
dependencies are deployed before their dependents).

For testing changes before deployment, there is the
`--update-schema-stage` option. This is currently set up to deploy to
the original dataset with the suffix `_test`.