Move metric definitions out into their own files
This is the first step to making it easier to define new metrics in
the future, separating the definitions from the contexts in which
they're used.

This also replaced the Metric.site_reports_field with a
Metric.condition(table) method, similar to MetricType, and defined
subclasses for the different metric types. This makes it easier to use
the metric definitions in metric-specific tables where the condition
is put in a WHERE clause rather than inline in the output field expression.