Only update views if they have changed
Compute a git-like hash of the view state, which is recorded in a
table in a metadata dataset. Then the remote data is updated in the
following circumstances:

* If there hasn't been an update yet today
* If the hash changed

The daily update means that we don't risk going out of sync with the
actual deployment due to changes made in prod; those will be
overwritten the next day.