Create a job to deploy views and routines (#403)
* Create a job to deploy views and routines

Each view or routine is defined by a meta.toml file containing the
name of the view or routine and other metadata, and a view.sql or
routine.sql file defining the actual SQL.

The SQL definition is a Jinja template, which has several useful
values and functions:

 * `ref()` for references to other data sets. This allows tracking of
 dependencies.

 * metrics, a dictionary of {metric_name: Metric}

 * metric_types, a list of MetricType objects

The routines are sorted topologically for deployment (so that
dependencies are deployed before their dependents).

For testing changes before deployment, there is the
`--update-schema-stage` option. This is currently set up to deploy to
the original dataset with the suffix `_test`.

* Add existing view and routine definitions to the repository.

* Only update views if they have changed

Compute a git-like hash of the view state, which is recorded in a
table in a metadata dataset. Then the remote data is updated in the
following circumstances:

* If there hasn't been an update yet today
* If the hash changed

The daily update means that we don't risk going out of sync with the
actual deployment due to changes made in prod; those will be
overwritten the next day.

* Require uv to run test script

It was broken without uv, so apparently that codepath wasn't being used anywhere.

* Add command to check templates

This ensures that they pass lints and that subsitution works, without guaranteeing that the output
is valid SQL.

* Initial tests for schema updates

* Fixes and additions to views sql

---------

Co-authored-by: Ksenia Berezina <kberezina@mozilla.com>
Co-authored-by: Anna Scholtz <anna@scholtzan.net>