Only consider outreach timed out if it was in progress (#414)
Previously if we had a contact date and a response date sufficiently
far in the past we'd consider outreach as `timed out` even without the
`webcompat:contact-in-progress` flag set. In practice this made it
impossible to get some bugs out of the `outreach` next action state
without deleting all record of contact.

This seems to be an incorrectly parenthesized query condition rather
than an intentional choice.