Support loading and updating table schema
This is needed because we have some tables where we want the table to
get additional columns as we add new metrics, and that's easier to do
if we can template the table definitions, as we do for views and
routines.

Tables are kept in `data/sql/<dataset>/tables/<name>`

Like views and routines they get a `meta.toml` file. They also get a
`table.toml` which is passed through jinja to produce the final table
definition.

The collection of all tables, views and routines is loaded into a
Project class at startup. This is then passed into jobs as part of the
Context object.

The update_schema job is updated to first create any newly added
tables, or add missing columns to existing tables. This happens before
views or routines are updated, since they may depend on tables, but
tables never depend on each other.

The Project object is also designed to allow referencing known
tables/views/routines for use in queries. For example one can write
something like:

```
client.query(f"SELECT * FROM `{project["webcompat_knowledge_base"]["bugzilla_bugs"]}`")
```

and it will automatically produce the correct reference given the
project id and value of the staging command line argument.