Move tool to update staging data into update_schema (#415)
This allows it to reuse the name mapping and view creation logic, and
so we only need custom code for cloning the tables.

Also instead of a hardcoded list, we query the original datasets for
the tables to create.

Note that this is currently broken with materialized views (which have
a `table_type` of `TABLE` rather than `VIEW`, so we currently just
ignore errors creating the test data).

Co-authored-by: Ksenia <kberezina@mozilla.com>