Convert the way that rescores work to match the way we store schema
Now that we have the schemas scored in-tree we don't want to directly
deploy tables for prospective rescores. Instead the approach we take
is to make the metric-rescore command do two things:

* Create templates for the post-rescore scored_site_reports, any
routines that need to be updated as part of the rescore, and a
template for a delta view that selects bugs that are changed by the
rescore.

* Replace: update the scored site reports templates with the new
version from a rescore, and move the current version to an archive
dataset, and similarly for routines modified by the rescore.

The idea is that one first deploys the new scored site reports, to
allow the change to be audited, and then deploys the change that
replaces the original scored site reports with the new one once the
rescore is ready to go.