Record rescores before updating schema.
If we deploy a rescore we want to first record the delta between the
old scores and the new scores. This is a bit tricky because we want
the update to the templates to go through the PR process and be
deployed by the ETL, but we want to record the rescores before we
actually update the schema.

To solve this we add an extra lifecycle point in the schema update for
tasks that run before deploying the schema. Here we call into
metric_rescore to check for a metric rescore that a) isn't marked as
staging and b) isn't deployed. If that's found we assume we're about
to deploy it and record the score updates resulting from that change
to the metric.

This isn't a perfect system since the changes aren't at all
atomic (e.g. the rescore could succeed, but the deploy fail), but it's
probably good enough for something we do relatively rarely.